mgmt: hawkbit: seperate header files
seperate the hawkbit header files, to make
it clearer.
