+++ MAINTAINERS.yml
@@ -2725,6 +2725,7 @@ hawkBit:
     - ycsin
   files:
     - subsys/mgmt/hawkbit/
+    - include/zephyr/mgmt/hawkbit/
     - include/zephyr/mgmt/hawkbit.h
     - samples/subsys/mgmt/hawkbit/
   labels:
